% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdy2mmy.R
\name{bdy2mmy}
\alias{bdy2mmy}
\title{Computing money market yield (MMY) for a T-bill}
\usage{
bdy2mmy(bdy, t)
}
\arguments{
\item{bdy}{bank discount yield}

\item{t}{number of days remaining until maturity}
}
\description{
Computing money market yield (MMY) for a T-bill
}
\examples{
bdy2mmy(bdy=0.045,t=120)
}
\seealso{
\code{\link{bdy}}
}

