% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash.ratio.R
\name{cash.ratio}
\alias{cash.ratio}
\title{cash ratio -- Liquidity ratios measure the firm's ability to satisfy its short-term obligations as they come due.}
\usage{
cash.ratio(cash, ms, cl)
}
\arguments{
\item{cash}{cash}

\item{ms}{marketable securities}

\item{cl}{current liabilities}
}
\description{
cash ratio -- Liquidity ratios measure the firm's ability to satisfy its short-term obligations as they come due.
}
\examples{
cash.ratio(cash=3000,ms=2000,cl=2000)
}
\seealso{
\code{\link{current.ratio}}

\code{\link{quick.ratio}}
}

