\name{coefficient.variation}
\alias{coefficient.variation}
\title{Computing Coefficient of variation}
\usage{
coefficient.variation(sd, avg)
}
\arguments{
  \item{sd}{standard deviation}

  \item{avg}{average value}
}
\description{
Computing Coefficient of variation
}
\examples{
coefficient.variation(sd=0.15,avg=0.39)
}
\seealso{
\code{\link{Sharpe.ratio}}
}

