\name{current.ratio}
\alias{current.ratio}
\title{current ratio -- Liquidity ratios measure the firm's ability to satisfy its short-term obligations as they come due.}
\usage{
current.ratio(ca, cl)
}
\arguments{
  \item{ca}{current assets}

  \item{cl}{current liabilities}
}
\description{
current ratio -- Liquidity ratios measure the firm's
ability to satisfy its short-term obligations as they come
due.
}
\examples{
current.ratio(ca=8000,cl=2000)
}
\seealso{
\code{\link{cash.ratio}}

\code{\link{quick.ratio}}
}

