% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt.ratio.R
\name{debt.ratio}
\alias{debt.ratio}
\title{debt ratio -- Solvency ratios measure the firm's ability to satisfy its long-term obligations.}
\usage{
debt.ratio(td, ta)
}
\arguments{
\item{td}{total debt}

\item{ta}{total assets}
}
\description{
debt ratio -- Solvency ratios measure the firm's ability to satisfy its long-term obligations.
}
\examples{
debt.ratio(td=6000,ta=20000)
}
\seealso{
\code{\link{total.d2e}}

\code{\link{lt.d2e}}

\code{\link{financial.leverage}}
}

