% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear.continuous.R
\name{ear.continuous}
\alias{ear.continuous}
\title{Convert stated annual rate to the effective annual rate with continuous compounding}
\usage{
ear.continuous(r)
}
\arguments{
\item{r}{stated annual rate}
}
\description{
Convert stated annual rate to the effective annual rate with continuous compounding
}
\examples{
ear.continuous(r=0.1)

ear.continuous(0.03)
}
\seealso{
\code{\link{ear}}

\code{\link{r.norminal}}
}

