% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear2bey.R
\name{ear2bey}
\alias{ear2bey}
\title{bond-equivalent yield (BEY), 2 x the semiannual discount rate}
\usage{
ear2bey(ear)
}
\arguments{
\item{ear}{effective annual rate}
}
\description{
bond-equivalent yield (BEY), 2 x the semiannual discount rate
}
\examples{
ear2bey(ear=0.08)
}
\seealso{
\code{\link{ear}}
}

