\name{ear2hpr}
\alias{ear2hpr}
\title{Computing HPR, the holding period return}
\usage{
  ear2hpr(ear, t)
}
\arguments{
  \item{ear}{effective annual rate}

  \item{t}{number of days remaining until maturity}
}
\description{
  Computing HPR, the holding period return
}
\examples{
ear2hpr(ear=0.05039,t=150)
}
\seealso{
  \code{\link{hpr2ear}}

  \code{\link{ear}}

  \code{\link{hpr}}
}

