% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.leverage.R
\name{financial.leverage}
\alias{financial.leverage}
\title{financial leverage -- Solvency ratios measure the firm's ability to satisfy its long-term obligations.}
\usage{
financial.leverage(te, ta)
}
\arguments{
\item{te}{total equity}

\item{ta}{total assets}
}
\description{
financial leverage -- Solvency ratios measure the firm's ability to satisfy its long-term obligations.
}
\examples{
financial.leverage(te=16000,ta=20000)
}
\seealso{
\code{\link{total.d2e}}

\code{\link{lt.d2e}}

\code{\link{debt.ratio}}
}

