% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.uneven.R
\name{fv.uneven}
\alias{fv.uneven}
\title{Computing the future value of an uneven cash flow series}
\usage{
fv.uneven(r, cf)
}
\arguments{
\item{r}{stated annual rate}

\item{cf}{uneven cash flow}
}
\description{
Computing the future value of an uneven cash flow series
}
\examples{
fv.uneven(r=0.1, cf=c(-1000, -500, 0, 4000, 3500, 2000))
}
\seealso{
\code{\link{fv.simple}}
}
