% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric mean return}
\usage{
geometric.mean(r)
}
\arguments{
\item{r}{returns over multiple periods}
}
\description{
Geometric mean return
}
\examples{
geometric.mean(r=c(-0.0934, 0.2345, 0.0892))
}
