% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.mean.R
\name{harmonic.mean}
\alias{harmonic.mean}
\title{harmonic mean, average price}
\usage{
harmonic.mean(p)
}
\arguments{
\item{p}{price over multiple periods}
}
\description{
harmonic mean, average price
}
\examples{
harmonic.mean(p=c(8,9,10))
}

