% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr2bey.R
\name{hpr2bey}
\alias{hpr2bey}
\title{bond-equivalent yield (BEY), 2 x the semiannual discount rate}
\usage{
hpr2bey(hpr, t)
}
\arguments{
\item{hpr}{holding period return}

\item{t}{number of month remaining until maturity}
}
\description{
bond-equivalent yield (BEY), 2 x the semiannual discount rate
}
\examples{
hpr2bey(hpr=0.02,t=3)
}
\seealso{
\code{\link{hpr}}
}
