% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr2mmy.R
\name{hpr2mmy}
\alias{hpr2mmy}
\title{Computing money market yield (MMY) for a T-bill}
\usage{
hpr2mmy(hpr, t)
}
\arguments{
\item{hpr}{holding period return}

\item{t}{number of days remaining until maturity}
}
\description{
Computing money market yield (MMY) for a T-bill
}
\examples{
hpr2mmy(hpr=0.01523,t=120)
}
\seealso{
\code{\link{hpr}}

\code{\link{mmy2hpr}}
}

