% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iss.R
\name{iss}
\alias{iss}
\title{calculate the net increase in common shares from the potential exercise of stock options or warrants}
\usage{
iss(amp, ep, n)
}
\arguments{
\item{amp}{average market price over the year}

\item{ep}{exercise price of the options or warrants}

\item{n}{number of common shares that the options and warrants can be convened into}
}
\description{
calculate the net increase in common shares from the potential exercise of stock options or warrants
}
\examples{
iss(amp=20,ep=15,n=10000)
}
\seealso{
\code{\link{diluted.EPS}}
}

