\name{lt.d2e}
\alias{lt.d2e}
\title{long-term debt-to-equity -- Solvency ratios measure the firm's ability to satisfy its long-term obligations.}
\usage{
  lt.d2e(ltd, te)
}
\arguments{
  \item{ltd}{long-term debt}

  \item{te}{total equity}
}
\description{
  long-term debt-to-equity -- Solvency ratios measure the
  firm's ability to satisfy its long-term obligations.
}
\examples{
lt.d2e(ltd=8000,te=20000)
}
\seealso{
  \code{\link{total.d2e}}

  \code{\link{debt.ratio}}

  \code{\link{financial.leverage}}
}

