\name{mmy2hpr}
\alias{mmy2hpr}
\title{Computing HPR, the holding period return}
\usage{
mmy2hpr(mmy, t)
}
\arguments{
  \item{mmy}{money market yield}

  \item{t}{number of days remaining until maturity}
}
\description{
Computing HPR, the holding period return
}
\examples{
mmy2hpr(mmy=0.04898,t=150)
}
\seealso{
\code{\link{bdy2mmy}}

\code{\link{hpr2mmy}}

\code{\link{hpr}}
}

