% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm}
\alias{npm}
\title{net profit margin -- Evaluate a company's financial performance}
\usage{
npm(ni, rv)
}
\arguments{
\item{ni}{net income}

\item{rv}{revenue (sales)}
}
\description{
net profit margin -- Evaluate a company's financial performance
}
\examples{
npm(ni=8000,rv=20000)
}
\seealso{
\code{\link{gpm}}
}

