% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.perpetuity.R
\name{pv.perpetuity}
\alias{pv.perpetuity}
\title{Estimate present value of a perpetuity}
\usage{
pv.perpetuity(r, pmt, g = 0, type = 0)
}
\arguments{
\item{r}{discount rate, or the interest rate at which the amount will be compounded each period}

\item{pmt}{payment per period}

\item{g}{growth rate of perpetuity}

\item{type}{payments occur at the end of each period (type=0); payments occur at the beginning of each period (type=1)}
}
\description{
Estimate present value of a perpetuity
}
\examples{
pv.perpetuity(r=0.1,pmt=1000,g=0.02)

pv.perpetuity(r=0.1,pmt=1000,type=1)

pv.perpetuity(r=0.1,pmt=1000) 
}
\seealso{
\code{\link{r.perpetuity}}
}
