\name{r.continuous}
\alias{r.continuous}
\title{Convert a given norminal rate to a continuous compounded rate}
\usage{
  r.continuous(r, m)
}
\arguments{
  \item{r}{norminal rate}

  \item{m}{number of times compounded each year}
}
\description{
  Convert a given norminal rate to a continuous compounded
  rate
}
\examples{
r.continuous(0.03,4)
}
\seealso{
  \code{\link{r.norminal}}
}

