% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.perpetuity.R
\name{r.perpetuity}
\alias{r.perpetuity}
\title{Rate of return for a perpetuity}
\usage{
r.perpetuity(pmt, pv)
}
\arguments{
\item{pmt}{payment per period}

\item{pv}{present value}
}
\description{
Rate of return for a perpetuity
}
\examples{
r.perpetuity(pmt=4.5,pv=-75) 
}
\seealso{
\code{\link{pv.perpetuity}}
}

