\name{Sharpe.ratio}
\alias{Sharpe.ratio}
\title{Computing Sharpe Ratio}
\usage{
Sharpe.ratio(rp, rf, sd)
}
\arguments{
  \item{rp}{portfolio return}

  \item{rf}{risk-free return}

  \item{sd}{standard deviation of portfolio retwns}
}
\description{
Computing Sharpe Ratio
}
\examples{
Sharpe.ratio(rp=0.038,rf=0.015,sd=0.07)
}
\seealso{
\code{\link{coefficient.variation}}

\code{\link{SFRatio}}
}

