\name{total.d2e}
\alias{total.d2e}
\title{total debt-to-equity -- Solvency ratios measure the firm's ability to satisfy its long-term obligations.}
\usage{
  total.d2e(td, te)
}
\arguments{
  \item{td}{total debt}

  \item{te}{total equity}
}
\description{
  total debt-to-equity -- Solvency ratios measure the
  firm's ability to satisfy its long-term obligations.
}
\examples{
total.d2e(td=6000,te=20000)
}
\seealso{
  \code{\link{total.d2e}}

  \code{\link{debt.ratio}}

  \code{\link{financial.leverage}}
}

