\name{wpr}
\alias{wpr}
\title{Weighted mean as a portfolio return}
\usage{
wpr(r, w)
}
\arguments{
  \item{r}{returns of the individual assets in the
  portfolio}

  \item{w}{corresponding weights associated with each of
  the individual assets}
}
\description{
Weighted mean as a portfolio return
}
\examples{
wpr(r=c(0.12, 0.07, 0.03),w=c(0.5,0.4,0.1))
}

