% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.rate.R
\name{discount.rate}
\alias{discount.rate}
\title{Computing the rate of return for each period}
\usage{
discount.rate(n, pv, fv, pmt, type = 0, lower = 1e-04, upper = 1)
}
\arguments{
\item{n}{number of periods}

\item{pv}{present value}

\item{fv}{future value}

\item{pmt}{payment per period}

\item{type}{payments occur at the end of each period (type=0); payments occur at the beginning of each period (type=1)}

\item{lower}{the lower end points of the rate of return to be searched.}

\item{upper}{the upper end points of the rate of return to be searched.}
}
\description{
Computing the rate of return for each period
}
\examples{
discount.rate(n=5,pv=0,fv=600,pmt=-100,type=0)
}
\seealso{
\code{\link{fv.simple}}

\code{\link{fv.annuity}}

\code{\link{fv}}

\code{\link{pv}}

\code{\link{pmt}}

\code{\link{n.period}}
}
