\name{irr}
\alias{irr}
\title{Computing IRR, the internal rate of return}
\usage{
  irr(cf)
}
\arguments{
  \item{cf}{cash flow,the first cash flow is the initial
  outlay}
}
\description{
  Computing IRR, the internal rate of return
}
\examples{
irr(cf=c(-5, 1.6, 2.4, 2.8))
}
\seealso{
  \code{\link{pv.uneven}}
}

