% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr2.R
\name{irr2}
\alias{irr2}
\title{Computing IRR, the internal rate of return}
\usage{
irr2(cf, cutoff = 0.01, from = -1, to = 1, step = 1e-06)
}
\arguments{
\item{cf}{cash flow,the first cash flow is the initial outlay}

\item{cutoff, }{threshold to take npv as zero}

\item{from, }{smallest irr to try}

\item{to, }{largest irr to try}

\item{step, }{increment of the irr}
}
\description{
Computing IRR, the internal rate of return
}
\examples{
# irr(cf=c(-5, 1.6, 2.4, 2.8))
}
\seealso{
\code{\link{irr}}
}

